integrating to moka via server

+++ database/tokens/token.table.sql
@@ -2,8 +2,8 @@ CREATE TABLE tokens (
 id UUID NOT NULL PRIMARY KEY,
 access_token TEXT NOT NULL,
 token_type VARCHAR(50) NOT NULL,
-expires_in INT NOT NULL,
-expires_at INT NOT NULL,
+expires_in BIGINT NOT NULL,
+expires_at BIGINT NOT NULL,
 scope TEXT NOT NULL,
 refresh_token TEXT NOT NULL,
 created_at TIMESTAMPTZ NOT NULL
